+++ jobs-user-service/src/main/java/com/example/jobsuserservice/dto/response/FileDBResponse2.java
@@ -4,7 +4,7 @@ package com.example.jobsuserservice.dto.response;
 public class FileDBResponse2 {
 
     private String cvskill;
-    private Float match;
+    private Double match;
     private String jdskill;
     private String id;
     private String name;
@@ -24,11 +24,11 @@ public class FileDBResponse2 {
         this.cvskill = cvskill;
     }
 
-    public Float getMatch() {
+    public Double getMatch() {
         return match;
     }
 
-    public void setMatch(Float match) {
+    public void setMatch(Double match) {
         this.match = match;
     }
